cromaticar-backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
asyncpg==0.29.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
googlesearch-python==1.2.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0